Locale

Description

Módulo de tradução do texto ao longo de todo o projeto.

Documentation

localeVersion :: Int

localeDirPath :: IO String

data LocaleId

Constructors

EN_US

PT_PT

getCachedLocale :: LocaleId -> IO Locale

getLocalizedString :: Locale -> String -> IO String

A função getLocalizedString, dado um Locale, traduz a chave de tradução para a linguagem do mesmo Locale.

getLocalizedStringP :: Locale -> String -> String

data Locale

Lista de locales disponíveis universalmente.

Representa um Locale.

Constructors

Locale

Fields

version :: Int

lang :: String

locales :: Locales

data Locales

Representa a lista de locales disponíveis e requeridos para um locale.

Constructors

Locales

Fields

menu_yes :: String

menu_no :: String

menu_resume :: String

menu_respawn :: String

menu_exit :: String

menu_play :: String

menu_options :: String

menu_options_english :: String

menu_options_portugues :: String

menu_save :: String

menu_save_save :: String

menu_save_save_verb :: String

menu_save_load_verb :: String

menu_save_delete_verb :: String

menu_save_delAll_conf :: String

menu_save_del_conf :: String

menu_save_createdAt :: String

menu_save_modifiedAt :: String

menu_save_lastModifiedAt :: String

menu_save_highscore :: String

menu_save_totalCoins :: String

menu_save_details :: String

menu_save_clear :: String

ensureStructure :: Locale -> Bool

loadLocale :: LocaleId -> IO Locale

A função loadLocale carrega um Locale a partir de um ficheiro.

en_US :: IO Locale

pt_PT :: IO Locale