Render.Backend.Atlas

Description

Módulo de leitura e processamento de Texture Atlas, permitindo que várias texturas se encontrem num único ficheiro de imagem, permitindo um uso mais eficiente da memória.

Um Texture Atlas necessita de ser uma imagem de 24-bits para evitar a utilização de um offset para as coordenadas.

Documentation

type Atlas = BitmapData

Representa um Texture Atlas.

data AtlasSection

Representa uma secção de um Texture Atlas.

Constructors

AtlasSection

Fields

(Int, Int)

Representa a posição (x,y) de uma secção. As posições são calculadas a partir do canto inferior esquerdo. Os valores de x e de y aumentam no sentido esquerda-direita e baixo-cima, respetivamente.

(Int, Int)

Representa o tamanho da secção. Segue as mesmas regras da posição.

pic2bmp :: Picture -> Atlas

Converte uma Picture num Atlas

loadAtlas :: String -> IO Atlas

Carrega um Atlas a partir de um ficheiro.

getAtlasSection :: Atlas -> AtlasSection -> Picture

Obtém uma secção de um Atlas. AVISO: Ver AtlasSection.

getAtlasSize :: Atlas -> (Int, Int)

Obtém o tamanho de um Atlas.

data OffsetEntry

Representa um Sprite num Atlas.

As posições são calculadas a partir do canto inferior esquerdo. Os valores de x e de y aumentam no sentido esquerda-direita e baixo-cima, respetivamente.

Constructors

OffsetEntry

Fields

x :: Int

Representa a posição x de um Sprite.

y :: Int

Representa a posição y de um Sprite.

w :: Int

Representa a largura de um Sprite.

h :: Int

Representa a altura de um Sprite.

type OffsetAtlas = Map String OffsetEntry

Representa os Sprites de um Atlas.

data Offsets

Representa uma tabela de Offsets de um Atlas.

Constructors

Offsets

Fields

atlas :: OffsetAtlas

parseAtlasSector :: Atlas -> OffsetAtlas -> Map String Picture

Transforma um mapa de Sprites em um mapa de Pictures