Render.Backend.Font

Description

Módulo de renderização de texto utilizando Fonts customizadas.

Documentation

type OffsetTable = [Int]

Representa os offsets de cada caractére. Valores negativos representam um shift á esquerda do texto e valores positivos representam shifts á direita do texto.

defaultOffsetTable :: OffsetTable

data FontData

Representa um Font no sistema de renderização.

Constructors

FontData

Fields

Atlas

O Texture Atlas da Font

(Int, Int)

O tamanho (x,y) base de cada caractere

OffsetTable

A Tabela de Offsets

makeFont :: String -> (Int, Int) -> IO FontData

Cria uma Font a partir de um Texture Atlas.

getFontCharPos :: FontData -> Char -> (Int, Int)

Obtém as coordenadas de um charactere na Font Especificada.

getCharFont :: FontData -> Char -> Picture

Obtém um caractere renderizado com a Font fornecida.

data TextAlignment

Representa o alinhamento de um texto multi-linha.

Constructors

AlignLeft

AlignCenter

AlignRight

strToPic :: FontData -> String -> TextAlignment -> Picture

Converte uma string para uma Picture contendo a mesma string renderizada com a Font fornecida.

stringSize :: FontData -> String -> (Int, Int)

Obtém o tamanho de uma String segundo uma Font.

trueStringSize :: FontData -> String -> (Float, Float)

Obtém o tamanho real de uma String segundo uma Font, representada numa Picture.