Render.Play

Description

Módulo de interface Jogador-Jogo. Permite a interação do Jogador com o Jogo e vice-versa. Contém funções que mudam o estado do Jogo mediante a interação do Jogador.

GameState

Estruturas

data GameState

A estrutura GameState define os dados de um jogo.

Mapa de Estrutura

JSON-notation + Haskell (type-notations + comments):

   {
       "flags": { -- Flags que modificam o jogo globalmente.
           "debug"   :: Bool, -- Indica se o jogo foi iniciado no modo debug.
           "godmode" :: Bool, -- Indica se o jogador possuí o GodMode ativado.
           "noclip"  :: Bool  -- Indica se o jogador possuí o NoClip ativado.
       },
       "state": { -- O estado global do jogo.
           "inGame" :: Bool,    -- Indica se o jogador se encontra num jogo.
           "inMenu" :: Bool,    -- Indica se o jogador se encontra num menu.
           "inGameMenu" :: Bool -- Indica se o jogador se encontra num menu dentro de um jogo.
       },
       "levelData": { -- O estado do nível atual.
           "mapType" :: MapType,  -- O tipo do mapa atual.
           "game" :: Jogo,        -- Os dados da grid do Mapa atual.
           "paused" :: Bool,      -- Indica se o jogador possuí o jogo pausado.
           "killed" :: Bool,      -- Indica se o jogador morreu.
           "playerData": {
               "points" :: Integer -- Os pontos obtidos na ronda atual.
           }
       },
       "globalPlayerData": { -- O estado global do jogador.
           "highScore" :: Integer -- A maior pontuação atingida pelo jogador.
       },
       "internal": { -- Cache interna
           "randomGen" :: [Int],                 -- O gerador de números aleatórios do jogo.
           "locale" :: Locale,                   -- O Locale atualmente usado.
           "logger" :: Logger                    -- O Logger atualmente usado.
           "keys" :: S.Set Key,                  -- O Set de teclas atualmente pressionadas.
           "clock" :: Clock,                     -- Os valores dos temporizadores utilizados.
           "textures" :: M.Map String G.Picture, -- A memória de Texturas atualmente carregadas.
           "fonts" :: M.Map String FontData      -- A memória de Fonts atualmente carregadas.
       }
   }

Constructors

GameState

Fields

flags :: Flags

state :: GSState

levelData :: Maybe LevelData

globalPlayerData :: GlobalPlayerData

internal :: InternalData

menu :: MenuData

data Flags

Constructors

Flags

Fields

debug :: Bool

godmode :: Bool

noclip :: Bool

data GSState

Constructors

GSState

Fields

inGame :: Bool

inMenu :: Bool

inGameMenu :: Bool

data LevelData

Constructors

LevelData

Fields

mapType :: MapType

game :: Game

paused :: Bool

killed :: Bool

playerData :: LocalPlayerData

data LocalPlayerData

Constructors

LocalPlayerData

Fields

points :: Int

data Map

Constructors

Map Int MapOffset [(Terreno, [Obstaculo])]

data Game

Constructors

Game Player Map

data GlobalPlayerData

Constructors

GlobalPlayerData

Fields

highScore :: Integer

data InternalData

Constructors

InternalData

Fields

randomGen :: RndGen

locale :: Locale

logger :: LoggerGen

keys :: Set Key

clock :: Clock

textures :: Map String Picture

fonts :: Map String FontData

Sinalizadores

data Flag

Constructors

Debug

GodMode

NoClip

data MapType

Constructors

REGULAR

CUSTOM

data Player

Constructors

Player MapOffset Velocity Coordinates

type Velocity = (Float, Float)

type Coordinates = (Float, Float)

type MapOffset = Float

data MenuId

Constructors

MID_MainMenu

MID_Options

MID_Save

MID_Save_DelAllConf

MID_Save_Active

MID_Paused

MID_Killed

data Clock

Constructors

Clock

Fields

clock1 :: Int

clock2 :: Float

clock3 :: Int

clock4 :: Int

Funções

toggleFlag :: GameState -> Flag -> GameState

Muda o estado de uma flag de e para ativada.

togglePlay :: GameState -> GameState

A função togglePlay alterna entre o menu e o jogo.

getLevelData :: GameState -> Maybe LevelData

Obtém os dados do nível do jogo, se o jogo estiver ativado.

setLevelData :: GameState -> LevelData -> GameState

Modifica os dados do nível do jogo.

resetLevelData :: GameState -> GameState

Elimina os dados do nível do jogo.

getHighScore :: GameState -> Integer

setHighScore

Arguments

:: GameState
O valor do GameState atual.

-> Integer
O novo valor do HighScore.

-> GameState
O valor do GameState modificado.

Modifica o High Score do jogador para o valor fornecido se for maior que o HS atual. Caso contrário retorna o GameState atual.

modifyPoints

Arguments

:: GameState
O valor do GameState atual.

-> Int
O número de pontos a adicionar/subtrair do jogador.

-> GameState
O valor do GameState modificado.

Modifica o valor dos pontos que o jogador possuí globalmente.

getRandom :: GameState -> (Int, GameState)

Obtém um número aleatório a partir do gerador presente na cache do GameState atual.

getLocale :: GameState -> Locale

A função getLocale obtém os dados do Locale atualmente carregado.

getLocaleId :: GameState -> LocaleId

A função getLocaleId obtém o identificador do Locale atualmente carregado.

changeLocale :: GameState -> LocaleId -> IO GameState

Muda o Locale do GameState.

addKey :: GameState -> Key -> GameState

Adiciona uma Key á lista de keypresses.

removeKey :: GameState -> Key -> GameState

Remove uma Key da lista de keypresses.

hasKey :: GameState -> Key -> Bool

Verifica se uma Key existe na lista de keypresses.

getClock :: GameState -> Clock

setClock :: GameState -> Clock -> GameState

getClock1Value :: GameState -> Int

setClock1Value :: GameState -> Int -> GameState

getClock2Value :: GameState -> Float

setClock2Value :: GameState -> Float -> GameState

getClock3Value :: GameState -> Int

setClock3Value :: GameState -> Int -> GameState

getClock4Value :: GameState -> Int

setClock4Value :: GameState -> Int -> GameState

addTextures :: GameState -> Map String Picture -> GameState

A função addTextures adiciona um mapa de Texturas á memória.

getTextures :: GameState -> Map String Picture

A função getTextures obtém as Texturas atualmente carregadas.

addFonts :: GameState -> Map String FontData -> GameState

A função addFonts adiciona um mapa de Fonts á memória.

getFonts :: GameState -> Map String FontData

A função getFonts obtém as Fonts atualmente carregadas.

getSelectedMenu :: GameState -> MenuId

A função getSelectedMenu obtém o Menu selecionado.

setSelectedMenu :: GameState -> MenuId -> GameState

A função setSelectedMenu modifica o Menu selecionado.

getSelectedMenuButton :: GameState -> Int

A função getSelectedMenuButton obtém o botão selecionado do Menu selecionado.

setSelectedMenuButton :: GameState -> Int -> GameState

A função setSelectedMenuButton modifica o botão selecionado do Menu selecionado.

getSelectedMenuData :: GameState -> MenuDescriptor

A função setSelectedMenuData obtém o valor do Menu selecionado.

setSelectedMenuData :: GameState -> MenuDescriptor -> GameState

A função setSelectedMenuData modifica o valor do Menu selecionado.

shiftNSelectedMenuButton :: GameState -> Int -> GameState

A função shiftNSelectedMenuButton seleciona o menu colocado a n posições do botão atualmente selecionado i. Caso n + i exceda o tamanho da lista de botões,l a posição será n + i - length l

getMenuStack :: GameState -> [MenuId]

A função getMenuStack obtém o valor do histórico de menus.

setMenuStack :: GameState -> [MenuId] -> GameState

A função setMenuStack muda o valor do histórico de menus.

backMenuStack :: GameState -> GameState

A função backMenuStack seleciona o menu anterior no histórico de menus.

popFromMenuStack :: GameState -> GameState

A função backMenuStack remove o último menu do histórico de menus.

selectMenu :: GameState -> MenuId -> GameState

A função selectaMenu seleciona um menu.

desselectMenu :: GameState -> GameState

A função desselectMenu seleciona o menu anteriormente selectionado.

isNoClip :: GameState -> Bool

Determina se o jogador encontra-se em NoClip.

isGodMode :: GameState -> Bool

Determina se o jogador encontra-se em GodMode.

Constantes

defaultGameState :: Int -> IO GameState

Valores iniciais do GameState.

defaultLevelData :: LevelData

Valores iniciais do LevelData.