Render.World.Level

Description

Módulo que contém as funções para o Jogo ativo.

Documentation

getGame :: GameState -> Maybe Game

Obtém os dados do Jogo atual.

getGamePlayer :: GameState -> Maybe Player

Obtém os dados do Jogador.

getPlayerData :: GameState -> Maybe LocalPlayerData

Obtém os dados relacionados com o Jogador neste nível.

isPaused :: GameState -> Bool

Determina se o jogo está pausado.

isDead :: GameState -> Bool

Determina se o Jogador morreu.

isInMenu :: GameState -> Bool

Determina se o Jogador se encontra num menu.

isInGame :: GameState -> Bool

Determina se o Jogador se encontra num jogo.

isInGameMenu :: GameState -> Bool

Determina se o Jogador encontra-se num menu dentro de um jogo.

getGameMap :: GameState -> Maybe Map

Obtém o Mapa atual do jogo.

getGameLine :: GameState -> Int -> Maybe Line

Obtém uma linha do Mapa atual.

getGameWidth :: GameState -> Maybe Int

Obtém a largura do Mapa atual.

getGamePlayerCoords :: GameState -> Maybe Coordinates

Obtém as coordenadas exatas do Jogador no mapa atual.

getGamePlayerOffset :: GameState -> Maybe MapOffset

Obtém o offset do Mapa atual.

getGamePlayerVelocity :: GameState -> Maybe Velocity

Obtém a velocidade do jogador.

getDisplayPlayerCoords :: GameState -> Maybe Coordenadas

Obtém as coordenadas base do Jogador.

getPlayerLevelPoints :: GameState -> Maybe Int

Obtém os pontos do Jogador no nível atual.

setGame :: GameState -> Game -> Maybe GameState

Modifica a estrutura Jogo para o nível atual.

setGameMap :: GameState -> Map -> Maybe GameState

Modifica o Mapa para o nível atual.

setPlayerData :: GameState -> LocalPlayerData -> Maybe GameState

Modifica os dados do Jogador no nível atual.

pauseGame :: GameState -> Maybe GameState

Pausa o jogo.

unpauseGame :: GameState -> Maybe GameState

Retoma o jogo.

killPlayer :: GameState -> Maybe GameState

Mata o Jogador no nível atual.

revivePlayer :: GameState -> Maybe GameState

Revive o Jogador no nível atual.

setPlayer :: GameState -> Player -> Maybe GameState

Modifica os dados do Jogador no nível atual.

setPlayerOffset :: GameState -> MapOffset -> Maybe GameState

Modifica o Offset do Jogador no nível atual.

setPlayerCoords :: GameState -> Coordinates -> Maybe GameState

Modifica as coordenadas do Jogador no nível atual.

setPlayerVelocity :: GameState -> Velocity -> Maybe GameState

Modifica a velocidade do Jogador no nível atual.

setPlayerLevelPoints :: GameState -> Int -> Maybe GameState

Modifica os pontos do Jogador no nível atual.

isNoClip :: GameState -> Bool

Determina se o jogador encontra-se em NoClip.

isGodMode :: GameState -> Bool

Determina se o jogador encontra-se em GodMode.