Tarefa1_2022li1g003

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2022/23.

Documentation

mapaValido

Arguments

:: Mapa
Um mapa a considerar para a verificação.

-> Bool
Resultado da verificação.

A função mapaValido verifica se um dado mapa pode ser, ou não, um mapa válido.

ensureTerreno

Arguments

:: Mapa
O Mapa original.

-> (Terreno, [Obstaculo])
A linha do mapa a ser verificada.

-> Bool
Resultado da verificação.

A função ensureTerreno recebe um Mapa e uma linha do Mapa, verificando se o comprimento da lista de obstáculos da linha corresponde exatamente à largura do Mapa.

AVISO: Esta é uma função interna, não desenhada para uso por funções externas a este ficheiro.

groupRios

Arguments

:: [(Terreno, [Obstaculo])]
Uma lista de linhas do Mapa.

-> [[(Terreno, [Obstaculo])]]
A lista com as linhas contíguas do tipo Rio.

A função groupRios agrupa as linhas contiguas do tipo Rio.

AVISO: Esta é uma função interna, não desenhada para uso por funções externas a este ficheiro.

groupEstradas

Arguments

:: [(Terreno, [Obstaculo])]
Uma lista de linhas do Mapa.

-> [[(Terreno, [Obstaculo])]]
A lista com as linhas contíguas do tipo Estrada.

A função groupEstradas agrupa as linhas contiguas do tipo Estrada.

AVISO: Esta é uma função interna, não desenhada para uso por funções externas a este ficheiro.

groupRelvas

Arguments

:: [(Terreno, [Obstaculo])]
Uma lista de linhas do Mapa.

-> [[(Terreno, [Obstaculo])]]
A lista com as linhas contíguas do tipo Relva.

A função groupRelvas agrupa as linhas contiguas do tipo Relva.

AVISO: Esta é uma função interna, não desenhada para uso por funções externas a este ficheiro.

ensureRios

Arguments

:: [(Terreno, b)]
Uma lista de linhas do Mapa.

-> Bool
Resultado da verificação.

A função ensureRios verifica se, numa lista de linhas do Mapa, caso existam linhas contíguas do tipo Rio, estas têm direções opostas.

AVISO: Esta é uma função interna, não desenhada para uso por funções externas a este ficheiro.

isRio

Arguments

:: (Terreno, [Obstaculo])
A linha a ser verificada.

-> Bool
Resultado da verificação.

A função isRio recebe uma linha do Mapa e verifica se esta possui as características necessárias para ser considerada um Rio.

AVISO: Esta é uma função interna, não desenhada para uso por funções externas a este ficheiro.

isEstrada

Arguments

:: (Terreno, [Obstaculo])
A linha a ser verificada.

-> Bool
Resultado da verificação.

A função isEstrada recebe uma linha do Mapa e verifica se esta possui as características necessárias para ser considerada uma Estrada.

AVISO: Esta é uma função interna, não desenhada para uso por funções externas a este ficheiro.

isRelva

Arguments

:: (Terreno, [Obstaculo])
A linha a ser verificada.

-> Bool
Resultado da verificação.

A função isRelva recebe uma linha do Mapa e verifica se esta possui as características necessárias para ser considerada uma Relva.

AVISO: Esta é uma função interna, não desenhada para uso por funções externas a este ficheiro.