Tarefa2_v2

Documentation

estendeMapa

Arguments

:: Map
O Mapa a ser estendido.

-> Int
Um número aleatório.

-> Int
Número a ser considerado.

-> Map
O mapa estendido.

A função estendeMapa gera e adiciona uma nova linha válida ao topo (i.e. primeira linha, visto de cima para baixo) de um dado mapa.

lineBuilder

Arguments

:: (Terreno, [Obstaculo])
A linha a ser criada.

-> (Terreno, [Obstaculo])
A linha anterior á linha a ser criada.

-> Int
Uma seed aleatória.

-> Int
Número a ser considerado.

-> ((Terreno, [Obstaculo]), Int)
Um tuplo contendo a linha gerada e a nova iteração do número aleatório fornecido.

proximosTerrenosValidos

Arguments

:: Map
O mapa sobre o qual avaliar os próximos terrenos válidos.

-> [Terreno]
A lista dos próximos terrenos válidos relativamente ao mapa fornecido.

A função proximosTerrenosValidos deve gerar a lista de terrenos passíveis de serem usados numa nova linha no topo do mapa dado.

proximosObstaculosValidos

Arguments

:: Int
A largura do mapa.

-> (Terreno, [Obstaculo])
A linha para a qual gerar obstáculos.

-> [Obstaculo]
Os tipos de obstáculos válidos para a linha.

A função proximosObstaculosValidos gera a lista de obstáculos passíveis de serem usados para continuar uma dada linha do mapa.