Tarefa3_2022li1g003

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2022/23.

Documentation

animaJogo

Arguments

:: Jogo
Jogo a ser considerado.

-> Jogada
Ação tomada pelo Jogador.

-> Jogo
Jogo atualizado após a Jogada.

A função animaJogo movimenta os obstáculos (de acordo com a Velocidade do terreno em que se encontram, e o personagem, de acordo com a Jogada dada.

Assume-se que o Mapa seja fornecido através de uma função futuramente implementada que garantirá que a posição inicial do Jogador é válida.

posicao

Arguments

:: Jogo
Jogo a ser considerado.

-> Jogada

-> Jogador
Posição atualizada após a Jogada.

A função posicao determina a posição em que ficará o Jogador após uma Jogada.

AVISO: Esta é uma função interna, não desenhada para uso por funções externas a este ficheiro.

isPositionValid

Arguments

:: Mapa
Mapa a ser considerado.

-> Coordenadas
Par de coordenadas a ser considerado.

-> Bool
Resultado da verificação.

A função isPositionValid determina se uma posição é válida, ou seja, se nessa posição não existe uma árvore ou um carro, ou se existe um tronco.

AVISO: Esta é uma função interna, não desenhada para uso por funções externas a este ficheiro.

isArvore

Arguments

:: (Terreno, [Obstaculo])
Linha a ser considerada.

-> Int
Coordenada x a ser considerada.

-> Bool
Resultado da verificação.

isAgua

Arguments

:: (Terreno, [Obstaculo])
Linha a ser considerada.

-> Int
Coordenada x a ser considerada.

-> Bool
Resultado da verificação.

isCarro

Arguments

:: (Terreno, [Obstaculo])
Linha a ser considerada.

-> Int
Coordenada x a ser considerada.

-> Bool
Resultado da verificação.

data Axis

Constructors

X

Y

attemptMove

Arguments

:: Jogo
Jogo a ser considerado.

-> Axis
Eixo a ser considerado.

-> Int
Quantidade de unidades a se mover no Eixo.

-> Jogo
Jogo atualizado após a Jogada.

moverObstaculos

Arguments

:: Jogador
O Jogador atual

-> (Terreno, [Obstaculo])
Linha de Mapa a ser considerada.

-> (Terreno, [Obstaculo])
Linha de Mapa com os objetos deslocados.

A função moverObstaculos desloca os obstáculos de uma linha consoante a velocidade do Terreno. Caso um obstáculo desapareça de um lado do mapa, este irá reaparecer no outro lado.

AVISO: Esta é uma função interna, não desenhada para uso por funções externas a este ficheiro.