Tarefa3_v2

Documentation

displayDimension :: (Int, Int)

Determina o tamanho da área do jogo

fps :: Float

Determina os FPS do jogo

animaJogo :: GameState -> GameState

A função animaJogo anima 1 frame do mapa do jogo.

animaJogador :: GameState -> GameState

A função animaJogador anima 1 frame do Jogador.

attemptMove :: GameState -> (Float, Float)

A função attemptMove limita os movimentos do Jogador a movimentos válidos.

isOnTronco :: GameState -> Bool

A função isOnTronco determina se o Jogador se encontra num tronco.

hasKeypresses :: GameState -> Bool

A função hasKeypresses determina se alguma key está pressionada.

resetKeypresses :: GameState -> GameState

A função hasKeypresses desativa todas as keys de movimento pressionadas.

animaMapa :: GameState -> GameState

A função animaMapa anima um 1 frame do Map.

moveMap :: Map -> Map

A função moveMap movimenta o Map em 1 frame.

moveObstacles :: Line -> MapOffset -> Line

A função moveObstacles movimenta os obstáculos do Map em 1 frame.

addPlayerStep :: Player -> MapOffset -> Player

A função addPlayerStep adiciona s ao Offset do Jogador.

addStep :: Map -> MapOffset -> Map

A função addStep adiciona s ao Offset do Map.

jogoTerminou

Arguments

:: GameState
Jogo a ser considerado.

-> GameState
Resultado da verificação.

isTronco

Arguments

:: (Terreno, [Obstaculo])
Linha a ser considerada.

-> Int
Coordenada x a ser considerada.

-> Bool
Resultado da verificação.

isEstrada

Arguments

:: (Terreno, [Obstaculo])
Linha a ser considerada.

-> Int
Coordenada x a ser considerada.

-> Bool
Resultado da verificação.

getRelativeTile :: (Terreno, [Obstaculo]) -> Int -> Int -> Obstaculo

A função getRelativeTile obtém o Obstáculo colocado a n casas de uma posição i, na mesma linha.

getTileSize :: Map -> (Float, Float)

A função getTileSize calcula o tamanho gráfico de uma tile.