Tarefa4_2022li1g014

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2022/23.

Para realizar a seguinte tarefa criamos uma função auxiliar que segue as condicionantes do enunciado, verificando se o jogador se encontra fora do mapa, na água, ou “debaixo”de um carro (i.e. na mesma posição de um carro.).

Documentation

atropelamento :: Jogador -> (Terreno, [Obstaculo]) -> Bool

carroemmovimento :: Coordenadas -> Int -> [Int] -> Bool

acomulador :: Eq a => a -> [a] -> [Int]

drowned :: Coordenadas -> (Terreno, [Obstaculo]) -> Bool

isGameOver :: Mapa -> Jogador -> Bool

foradomapa :: Jogador -> Mapa -> Bool

jogoTerminou :: Jogo -> Bool