Tarefa5_v2

Description

Módulo para a realização da reestruturação da Tarefa5, dado novo tipo de Jogo (Game).

Documentation

deslizaGame

Arguments

:: Int
Número a ser considerado.

-> Int
Número a ser considerado.

-> Game
Jogo a ser considerado.

-> Game
Jogo com Mapa e Coordenadas do Jogador atualizadas.

A função deslizaGame gera uma linha no topo do mapa, removendo a última, assim mantendo a quantidade de linhas de mapa.