Util

Description

Módulo contendo funções utilitárias gerais para a realização do projeto de LI1 em 2022/23.

Documentation

filterIndexed

Arguments

:: (a -> Int -> Bool)
A função de filtro, que toma um tuplo contendo um elemento e o seu índice na lista e retorna um valor booleano.

-> [a]
A lista a ser filtrada.

-> [(a, Int)]
A lista contendo tuplos com o elemento filtrado e o seu índice, respetivamente, para todos os elementos originais.

A função filterIndexed filtra os elementos de uma lista de acordo com um predicado fornecido pelo utilizador, o qual, recebendo cada elemento da lista seguido do seu índice na mesma lista deverá retornar o valor booleano True para elementos a receber e o valor booleano False para elementos a descartar.

A função retornará uma lista de tuplos contendo o elemento e o seu índice, respetivamente.

mapIndexed

Arguments

:: ((a, Int) -> b)
A função de mapeamento, que toma um tuplo contendo um elemento e o seu índice na lista e retorna o elemento mapeado.

-> [a]
A lista a ser mapeada.

-> [(b, Int)]
A lista contendo tuplos com o elemento mapeado e o seu índice, respetivamente, para todos os elementos originais.

A função mapIndexed mapeia os elementos de uma lista de acordo com um predicado fornecido pelo utilizador, o qual, recebendo cada elemento da lista seguido do seu índice na mesma lista deverá retornar o valor mapeado do elemento.

A função retornará uma lista de tuplos contendo o elemento mapeado e o seu índice, respetivamente.

anyIndexed

Arguments

:: ((a, Int) -> Bool)
A função de mapeamento, que toma um tuplo contendo um elemento e o seu índice na lista e retorna o elemento mapeado.

-> [a]
A lista a ser mapeada.

-> Bool
A lista contendo tuplos com o elemento mapeado e o seu índice, respetivamente, para todos os elementos originais.

A função anyIndexed verifica se pelo menos um elemento de uma lista satisfaz um predicado fornecido pelo utilizador, o qual, recbendo cada elemento da lista seguido do seu índice na mesma lista deverá retornar um valor booleano.

A função retornará um valor booleano.

mapWithCacheR

Arguments

:: (a -> b -> [c] -> (a, c))
A função de mapeamento.

-> a
O valor inicial da cache.

-> [b]
A lista a ser mapeada.

-> ([c], a)
A lista de elementos mapeados.

A função mapWithCacheR é a versão non-overloaded da função mapWithCache. A única diferença entre as mesmas é o valor de retorno, retornando esta função um tuplo contendo os elementos mapeados e o último valor da cache, respetivamente.

mapWithCache

Arguments

:: a
O valor inicial da cache.

-> (a -> b -> [c] -> (a, c))
A função de mapeamento.

-> [b]
A lista a ser mapeada.

-> [c]
A lista de elementos mapeados.

A função mapWithCache executa o mesmo comportamento da função predefinida map, com a exceção de que permite que um valor de cache c seja utilizado na próxima iteração.

Exemplos de Utilização

>>> mapWithCache 0 (\x c -> (x + c,c + 1)) [0,0,0]
[0,1,2]

getPRndElem

Arguments

:: [a]
Uma lista de elementos da qual selecionar um elemento aleatório. -> Int -- ^ max: O valor máximo possível para r.

-> Int
r: Um número pseudo-aleatório.

-> a
O elemento selecionado.

A função getPRndElem retorna um elemento pseudo-aleatório de uma lista a partir de um número pseudo-aleatório r pertencente ao intervalo [0,max].

getTimeSA :: IO Integer

A função getTimeSA retorna a data no momento da chamada no formato Unix Date Format (em segundos).

getDate :: IO (Int, Int, Integer)

A função getDate retorna a data no momento da chamada no formato DDMMYYYY.

formatDate :: UTCTime -> String

makeTime :: Integer -> UTCTime

scrambleNum

Arguments

:: Integer
O número a modificar.

-> Integer
O número resultante.

A função scrambleNum aplica uma série de operações matemáticas que transformam um número fornecido em um outro número sem significado.

AVISO: Esta é uma função interna, não desenhada para uso por funções externas a este ficheiro.

getPRand :: Int -> [Int]

A função getPRand gera um número pseudoaleatório a partir de uma seed fornecida pelo utilizador.

Para garantir uma sequência de números pseudoaleatórios, é recomendado que se use o resultado anterior da execução desta função.

shiftN :: Int -> [a] -> [a]

A função shiftN movimenta os elementos de uma lista num modo wrap-around.

Valores de n positivos movimentam os elementos para a direita n índices. Valores de n negativos movimentam os elementos para a esquerda n índices.

shiftNCI :: ([(a, Int)] -> Int) -> Int -> [a] -> [(a, Int)]

A função shiftNC movimenta os elementos de uma lista num modo wrap-around consoante o valor de um predicado fornecido pelo utilizador, que recebe o elemento a ser rodado e o seu índice na lista e que retorna: - 0 se o elemento passa o predicado. - 1 se o elemento falha o predicado, retornando a lista construida até então. - Qualquer outro valor se o elemento falha o predicado e uma lista vazia é retornada.

getTerrenoConstr :: Num a => Terreno -> a

A função getTerrenoConstr retorna um Int representando o identificador do tipo de terreno fornecido.

Indentificadores de Terrenos

0: Rio

1: Estrada

2: Relva

getTerrenoVel :: Terreno -> Velocidade

A função getTerrenoVel retorna a velocidade de um terreno, se for dinâmico (Rio,Estrada), ou 0 se o terreno for estático (Relva).

throws :: IO a -> IO Bool

Retorna um valor booleano verdadeiro se uma ação retorna uma exceção, ou um valor boolean falso se nenhuma exceção é retornada.

Adaptado de: https://stackoverflow.com/a/41127693

replaceAt :: [a] -> Int -> a -> [a]

replaceAtIndexes :: [a] -> [(Int, a)] -> [a]

wordsWhen :: (Char -> Bool) -> String -> [String]

withinBoundsWA :: Integral a => a -> a -> a -> a

withinBounds :: Integral a => a -> a -> a -> a

trimString :: [Char] -> Int -> Bool -> String

padString :: String -> Char -> Int -> String

capitalizeWord :: String -> String

getDecimal :: RealFrac b => b -> b

Retorna a parte decimal de um número.

fst3 :: (a, b, c) -> a

snd3 :: (a, b, c) -> b

thrd3 :: (a, b, c) -> c