Util.Logger

Description

Módulo de logging para mensagens.

Documentation

logDirPath :: IO String

defaultLogData :: LogData

defaultLogColorData :: LogColorData

data Level

Constructors

Info

Warn

Error

Debug

data LogData

Constructors

LogData

Fields

colors :: LogColorData

colored :: Bool

data LogColorData

Constructors

LogColorData

Fields

info :: Color

warn :: Color

error :: Color

debug :: Color

type Logger

Arguments

= String

-> IO ()
Representa uma instância de um Logger de Nível

type LoggerGen

Arguments

= Level

-> Logger
Representa uma instância de um Logger

prepareLog :: LogData -> Level -> String -> IO String

A função prepareLog, recebendo uma configuração de log, um nível de log e uma mensagem, retorna uma mensagem de log formatada pronta para ser utilizada pelas funções de logging.

AVISO: Esta é uma função interna, não desenhada para uso por funções externas a este ficheiro.

loggerFS :: LogData -> String -> LoggerGen

A função loggerFS, recebendo uma configuração de logs e um nome de ficheiro de log, retorna uma função que, dado o nível do log e a sua mensagem, executa o log para o ficheiro de log especificado.

loggerSB :: LogData -> LoggerGen

A função loggerSB, recebendo uma configuração de log, retorna uma função que, dado o nível do log e a sua mensagem, executa o log para o stdout.

loggerMixed :: LogData -> String -> LoggerGen

A função loggerMixed, recebendo uma configuração de logs e um nome de ficheiro de log, retorna uma função que, dado o nível do log e a sua mensagem, executa o log tanto para o stdout tanto como para o ficheiro de log especificado.

makeLogs :: LoggerGen -> (Logger, Logger, Logger, Logger)

A função makeLogs, recebendo um Logger parcialmente aplicado a dois argumentos do seu término, retorna um tuplo de nível 4 contendo Loggers derivados do Logger original parcialmente aplicados a cada nível de log (info, warn, error, debug).