Util.Logger.Colors

Description

Módulo de manipulação de mensagens de log, permitindo que cores sejam aplicadas ás mesmas.

Documentation

toANSI :: (Monoid m, IsString m) => m -> m

Converte um número em um código ANSI.

toANSI2 :: (Monoid t1, Monoid t, IsString t1, IsString t) => (t, t1) -> (t, t1)

Transforma um par de números num par de códigos ANSI.

makeANSIPair :: (Monoid m, IsString m) => (m, m) -> m -> m

Gera um par de cores para o background e o foreground do texto na consola.

type Color = (Monoid String, IsString String) => String -> String

newtype UseColor a

Constructors

Color a

reset :: Color

bold :: Color

dim :: Color

italic :: Color

underline :: Color

inverse :: Color

hidden :: Color

strikethrough :: Color

black :: Color

red :: Color

green :: Color

yellow :: Color

blue :: Color

magenta :: Color

cyan :: Color

white :: Color

gray :: Color

bgBlack :: Color

bgRed :: Color

bgGreen :: Color

bgYellow :: Color

bgBlue :: Color

bgMagenta :: Color

bgCyan :: Color

bgWhite :: Color