Util.Record

Description

Módulo de manipulação de Records. Permite obter e modificar propriedades de Records dinâmicamente.

Documentation

newtype T x

Representa um tipo genérico de transformação.

Constructors

T

Fields

unT :: x -> x

extQ :: (Typeable a, Typeable b) => (a -> q) -> (b -> q) -> a -> q

Extende uma query genérica para uma query específica.

extT :: (Typeable a, Typeable b) => (a -> a) -> (b -> b) -> a -> a

Extende uma transformação genérica para um caso específico a um tipo.

ext0 :: (Typeable a, Typeable b) => c a -> c b -> c a

Extensão de tipo extensível.

data Ti a

Representa uma Transformação Inteira.

Constructors

Ti Int a

gmapTi :: Data a => Int -> (forall b. Data b => b -> b) -> a -> a

Mapeia uma estrutura segundo uma Transformação Inteira.

fieldNames :: Data r => r -> [String]

Obtém as chaves das entradas de dados num Record.

fieldIndex :: Data r => String -> r -> Int

Garante que uma String é uma chave de entrada de um Record.

modifyField :: (Data r, Typeable v) => String -> (v -> v) -> r -> r

Modifica o valor de uma propriedade de um Record.

setField :: (Data r, Typeable v) => String -> v -> r -> r

Modifica uma propriedade de um Record.

getField :: (Data r, Typeable v) => String -> r -> v

Obtém uma propriedade de um Record. Necesita de uma assinatura de tipo estrita.